shubham/misc/requirements_streamlit.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
opencv-python>=4.8.0
pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pytz>=2023.3 